+++ packages/base-styles/_z-index.scss
@@ -187,6 +187,7 @@ $z-layers: (
 	".edit-site-layout__header": 2,
 	".edit-site-page-header": 2,
 	".edit-site-patterns__header": 2,
+	".edit-site-patterns__grid-pagination": 2,
 	".edit-site-layout__canvas-container": 2,
 	".edit-site-layout__sidebar": 1,
 	".edit-site-layout__canvas-container.is-resizing::after": 100,
